consolidated cloud search functions to one file
